agent/requirements.txt
langchain
langchain-community
langchain-core
langchain-text-splitters
google-genai
langchain-google-genai
langchain-huggingface
huggingface-hub
sentence-transformers
python-dotenv
django
pymupdf
openai